moved epoch advancers to hpp and cpp files.
