+++ src/persist/EpochSys.hpp
@@ -139,144 +139,9 @@ public:
 #include "PerThreadContainers.hpp"
 #include "ToBePersistedContainers.hpp"
 #include "ToBeFreedContainers.hpp"
+#include "EpochAdvancers.hpp"
 
 class EpochSys{
-    
-    /////////////////////
-    // Epoch Advancers //
-    /////////////////////
-
-    class EpochAdvancer{
-    public:
-        virtual void set_epoch_freq(int epoch_freq) = 0;
-        virtual void set_help_freq(int help_freq) = 0;
-        virtual void on_end_transaction(EpochSys* esys, uint64_t c) = 0;
-        virtual ~EpochAdvancer(){}
-    };
-
-    class SingleThreadEpochAdvancer : public EpochAdvancer{
-        // uint64_t trans_cnt;
-        padded<uint64_t>* trans_cnts;
-        uint64_t epoch_threshold = 0x1ULL << 19;
-        uint64_t help_threshold = 0x1ULL << 6;
-    public:
-        SingleThreadEpochAdvancer(GlobalTestConfig* gtc){
-            trans_cnts = new padded<uint64_t>[gtc->task_num];
-            for (int i = 0; i < gtc->task_num; i++){
-                trans_cnts[i].ui = 0;
-            }
-        }
-        void set_epoch_freq(int epoch_power){
-            epoch_threshold = 0x1ULL << epoch_power;
-        }
-        void set_help_freq(int help_power){
-            help_threshold = 0x1ULL << help_power;
-        }
-        void on_end_transaction(EpochSys* esys, uint64_t c){
-            assert(_tid != -1);
-            trans_cnts[_tid].ui++;
-            if (_tid == 0){
-                // only a single thread can advance epochs.
-                if (trans_cnts[_tid].ui % epoch_threshold == 0){
-                    esys->advance_epoch(c);
-                } 
-                // else if (trans_cnts[_tid].ui % help_threshold == 0){
-                //     esys->help_local();
-                // }
-            }
-            // else {
-            //     if (trans_cnts[_tid].ui % help_threshold == 0){
-            //         esys->help_local();
-            //     }
-            // }
-        }
-    };
-
-    class GlobalCounterEpochAdvancer : public EpochAdvancer{
-        std::atomic<uint64_t> trans_cnt;
-        uint64_t epoch_threshold = 0x1ULL << 14;
-        uint64_t help_threshold = 0x1ULL;
-    public:
-        // GlobalCounterEpochAdvancer();
-        void set_epoch_freq(int epoch_power){
-            epoch_threshold = 0x1ULL << epoch_power;
-        }
-        void set_help_freq(int help_power){
-            help_threshold = 0x1ULL << help_power;
-        }
-        void on_end_transaction(EpochSys* esys, uint64_t c){
-            uint64_t curr_cnt = trans_cnt.fetch_add(1, std::memory_order_acq_rel);
-            if (curr_cnt % epoch_threshold == 0){
-                esys->advance_epoch(c);
-            } else if (curr_cnt % help_threshold == 0){
-                esys->help_local();
-            }
-        }
-    };
-
-    class DedicatedEpochAdvancer : public EpochAdvancer{
-        EpochSys* esys;
-        std::thread advancer_thread;
-        std::atomic<bool> started;
-        uint64_t epoch_length = 100*1000;
-        void advancer(){
-            while(!started.load()){}
-            while(started.load()){
-                esys->advance_epoch_dedicated();
-                std::this_thread::sleep_for(std::chrono::microseconds(epoch_length));
-            }
-            // std::cout<<"advancer_thread terminating..."<<std::endl;
-        }
-    public:
-        DedicatedEpochAdvancer(GlobalTestConfig* gtc, EpochSys* es):esys(es){
-            if (gtc->checkEnv("EpochLength")){
-                epoch_length = stoi(gtc->getEnv("EpochLength"));
-            } else {
-                epoch_length = 100*1000;
-            }
-            if (gtc->checkEnv("EpochLengthUnit")){
-                std::string env_unit = gtc->getEnv("EpochLengthUnit");
-                if (env_unit == "Second"){
-                    epoch_length *= 1000000;
-                } else if (env_unit == "Millisecond"){
-                    epoch_length *= 1000;
-                } else if (env_unit == "Microsecond"){
-                    // do nothing.
-                } else {
-                    errexit("time unit not supported.");
-                }
-            }
-
-            started.store(false);
-            advancer_thread = std::move(std::thread(&DedicatedEpochAdvancer::advancer, this));
-            started.store(true);
-        }
-        ~DedicatedEpochAdvancer(){
-            // std::cout<<"terminating advancer_thread"<<std::endl;
-            started.store(false);
-            if (advancer_thread.joinable()){
-                advancer_thread.join();
-            }
-            // std::cout<<"terminated advancer_thread"<<std::endl;
-        }
-        void set_epoch_freq(int epoch_interval){
-        }
-        void set_help_freq(int help_interval){
-        }
-        void on_end_transaction(EpochSys* esys, uint64_t c){
-            // do nothing here.
-        }
-    };
-
-    class NoEpochAdvancer : public EpochAdvancer{
-        // an epoch advancer that does absolutely nothing.
-    public:
-        // GlobalCounterEpochAdvancer();
-        void set_epoch_freq(int epoch_power){}
-        void set_help_freq(int help_power){}
-        void on_end_transaction(EpochSys* esys, uint64_t c){}
-    };
-
     
 private:
     // persistent fields:
